config classes moved to the application/config folder
