+++ src/main/java/co/com/cargomaster/cargomaster/application/config/UseCasesConfig.java
@@ -1,4 +1,4 @@
-package co.com.cargomaster.cargomaster.applicationconfig;
+package co.com.cargomaster.cargomaster.application.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
